requirements.txt
streamlit
markdown
google-genai==1.24.0